requirements.txt
pytorch==2.0.1
pytorch-lightning==2.0.2
torch-geometric==2.3.0
transformers==4.33.1
evaluate==0.4.0
wandb==0.15.3
peft==0.5.0
adapters==0.1.1